i THINK sellLootbox is working. No working test on it yet

+++ backend/src/main/java/no/enterprise/exam/backend/entity/Users.java
@@ -62,7 +62,7 @@ public class Users {
         this.lootboxes = lootboxes;
     }
 
-    public long getCurrency() {
+    public int getCurrency() {
         return currency;
     }
